feat: :sparkles: Discord logging
With the updated rdrp_logs you can now log all actions that each admin does with rdrp_logs.
Make sure to add your webhookURN in the config file of rdrp_logs.
You can also log the admin actions not just in a discord chat but in the server console as well !!!

 p.s. The script will need rdrp_logs installed to work so be sure to install that right away!!

Requires rdrp_logs to work!!